requirements.txt
Flask
psycopg2-binary
SQLAlchemy
passlib
pyjwt